Remove starting application installation step
In Elixir >= 1.4 we no longer need to add the application because it is
configured in the dependency.

Even knowing of the change to Elixir 1.4 I got caught up on this because
I was focused on getting wifi setup. Since I added applications to my
application config I ran into other modules not working on my device. In
my host config I didn't have that issue so the problem was further
exacerbated by the modules being available on the host iex session.

Amos King @adkron amos@binarynoggin.com